requirements.txt
numpy
matplotlib
joblib
cvxpy
pandas
pyyaml
seaborn
scipy